==> Step1 Data Wrangling.ipynb <==
# Import Packages 

%matplotlib inline
#Import package for pandas dataframe
import pandas as pd
# Import the regular expression module
import re

# Import Dataset

r = pd.read_csv( 'ratings.csv' ) # ratings for different books
b = pd.read_csv( 'books.csv' ) # list of books and necessary infor about the books
t = pd.read_csv( 'tags.csv')  # Tag id and Tag names  used by readers to shelve a book
bt = pd.read_csv( 'book_tags.csv') #Records of tags received by different books 
tr = pd.read_csv( 'to_read.csv')# books marked by users as to read 


# Exploring the Datasets

## ratings.csv

# size of the dataset 
len(r)

# What kind of information does it have 
r.head()

# Check for data type, null entries
r.info()

# Any null entries
r.isnull().values.any()

# Check for duplicates
r.loc[r.duplicated(keep = False),:]

## books.csv

# size of the dataset 
len(b)

# What kind of information does it have 
b.head()

# Check for data type, null entries
b.info()

# Check for duplicates
b.loc[b.duplicated(keep = False),:]

b.isnull().values.any()

## tags.csv

# size of the dataset 
len(t)

# What kind of information does it have 
t.head()

# Check for data type, null entries
t.info()

# Check for duplicates
t.loc[t.duplicated(keep = False),:]

t.isnull().values.any()

## book_tags.csv

# size of the dataset 
len(bt)

# What kind of information does it have 
bt.head()

# Check for data type, null entries
bt.info()

# Check for duplicates
bt.loc[bt.duplicated(keep = False),:]

## to_read.csv

# size of the dataset 
len(tr)

# What kind of information does it have 
tr.head()

# Check for data type, null entries
tr.info()

# Check for duplicates
tr.loc[tr.duplicated(keep = False),:]

## Raw Dataset Initial analysis

> **ratings.csv** (r) is clean, it does not have any missing\null entries or duplicate rows.<br>
> **books.csv (b)** has some entires missing for some of the columns (i.e. isbn,isbn13, original_publication_year, original_title,language_code) <br>                                                   
> **tags.csv (t)** will require further investigation as the tag_names are not clear from the initial inspection. However, The dataset does not have any duplicate\missing entries <br>
> **book_tags.csv** (bt) has duplicate rows <br>
> **to_read.csv** (tr) looks clean <br>

## Merging datasets


###  'ratings.csv' & 'books.csv':

> Ratings are sorted chronologically, oldest first.

# Size of the dataset
len(r)

# Size of the dataset
len(b)

#Lets check if the "book_id" column has same elements 
if sorted(set(b.book_id)) == sorted(set(r.book_id)):
    print("The book_id column contains same elements, so it can be used to connect the two datasets")

# Lets check the number of unique elements for all features in ratings.csv dataset
r.nunique()

> There are 53424 users rating 10000 books in 5 categories from 1-5, so we can conclude that each book received multiple ratings

###   'tags.csv','book_tags.csv' &  'books.csv',:

> From the observation of these three dataframes, it seems the column 'tag_id' is common between dataframes for tags.csv and book_tags.csv. Also, the column, goodreads_book_id is common for dataframes between book_tags.csv and books.csv

> Lets take a look at more rows in tags.csv as from the few rows observed in section 1.1 it was unclear clear what kind of information the column contains

t.tail(10)

### Drop Duplicates
From our data inspection in section 4, we know that book_tags.csv has duplicates. 

bt = bt.drop_duplicates()

### Merge the Datasets

bt = bt.merge( t, on = 'tag_id' )

bt.head()

tag_table = bt.merge( b[[ 'goodreads_book_id', 'title']], on = 'goodreads_book_id' )
tag_table.head()

The combined table above contains all the tag_names that have been used by different users to catgorize the books. In the next step, we will perform some data cleaning to identify the most frequently occuring tag_names that can be used to categorize the books.

# Check the number of unique tag_names in the dataset prior filtering out non-ascii elements
tag_table.tag_name.nunique()

#Lets filter out rows with non - ascii characters
non_ascii_tags = tag_table[tag_table['tag_name'].str.contains(r'[^\x00-\x7F]+')]
tag_table =tag_table[~tag_table['tag_name'].str.contains(r'[^\x00-\x7F]+')]

#Lets check the size after applying the filter 
tag_table.tag_name.nunique()

#Lets filter out rows with non - ascii characters from the title
non_ascii_title = tag_table[tag_table['title'].str.contains(r'[^\x00-\x7F]+')]
tag_table =tag_table[~tag_table['title'].str.contains(r'[^\x00-\x7F]+')]

#Lets check the size after applying the filter 
tag_table.title.nunique()

### Remove Barely Used Tag_Names

# lets check the statistics for frequency of these tags in the book_id dataset
tag_table.tag_name.value_counts(dropna=False).describe()

> Based on the information above, it seems that there are now 32963 tag_names to categorize a total of 10000 books, but 75% of the tag_names have been used only 5 times by the all the 53424 users. These tag_names probably represent customized categories made by the user and will not be useful to cluster the books by popular categories. So lets try to identify tags that are widely used to shelf the books. To idetify such tags, we can group the dataframe ('tag_table') by tag_ids and count the number of books that each tag_id covered.

def popular_tags (tagname,M):
    if tagname in List:
        return tagname

tag_table_shortened = tag_table[['goodreads_book_id','tag_id','tag_name']]
books_per_tag = tag_table_shortened.groupby(['tag_id','tag_name']).count().sort_values(by='goodreads_book_id',ascending = False)
books_per_tag.rename(columns = {'goodreads_book_id':'Number of Books'}, inplace = True)
books_per_tag.head()

> It seems that the widely used tags (to-read, favorites, owned etc.) cover almost all the books. From this table, lets idetify and eliminate tags has been used for lesst than 300 books as they represent a negligible portion (only 3%) of the 10000 books in  collection

tags_to_keep = books_per_tag.index[books_per_tag['Number of Books']>300].tolist()
tag_id_to_keep, tag_names_to_keep = zip(*tags_to_keep)
len(set(tag_names_to_keep))
len(set(tag_id_to_keep))

# Define a function to create a Boolean Mask 
def Filtered(dataframe,List):
    if dataframe in List:
        dataframe = True
    else:
        dataframe = False
    return dataframe
# Apply the function defined above to make a boolean mask based on tag_id_to_keep 
Filter = tag_table.tag_id.apply(lambda x:Filtered(x,tag_id_to_keep ))
# Use the boolean mask to filter data
tag_table = tag_table[Filter]
# Use the boolean mask to filter data
tag_table.nunique()

## Define Some Generalized Categories by Clustering User Provided Tag_Names

### Tag for Favorite Books

Lets look at some of the remaining tag_names to checke if we can do further processing on the dataset.

tag_table.tag_name[50:100]

prog = re.compile('\w*[-]?[Ff]av\w*')
Favorites = []
for row in tag_table.tag_name[50:200]:
    result = prog.match(row)
    if bool(result) and row not in Favorites:
        Favorites.append(row)

Favorites

> It looks like the users have shelved their favorite books with tag_names such as ['my-favorites','faves','favorite-series', 'shelfari-favorites', 'favourite', 'childhood-favorites','favourite-books','favs'] etc. We will name all these categories as one common category favorites. <br>

> As a first step, lets define  functions to find custom categories following a similar pattern (e.g. (strings containing the word 'fav') and replace them with a common category [e.g. Favorites] in the 'tag_name' column.

def identify_custom_categories (tagname, pattern):
    prog = re.compile(pattern)
    matched_tags = []
    for row in tagname:
        result = prog.match(row)
        if bool(result) and row not in matched_tags:
            matched_tags.append(row)
    return matched_tags

def replace_custom_categories(tag_name, custom_names, preferred_category_name):
    if tag_name in custom_names:
        tag_name = preferred_category_name
    return tag_name

# Identify custom tag_names for favorite
pattern_for_Favorites = '\w*[-]?\w*[-]?[Ff]av\w*'
custom_tags_for_favorites = identify_custom_categories(tag_table.tag_name, str(pattern_for_Favorites))


# Rename the custom tag_names as Favorite
preferred_name = 'Favorite'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_favorites,preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos  = {}
KW_CategorY_Repos ['Category'] = ['Favorite']
KW_CategorY_Repos ['Possible Search KW'] = []
KW_CategorY_Repos ['Possible Search KW'].append(custom_tags_for_favorites)

# Check the size after filtering
tag_table.tag_name.nunique()
KW_CategorY_Repos

# Check if Replacement Worked
tag_table.tag_name [50:100]

## Tag for Children Books

Lets group tag names such as kids-books, childer-s-literature into one single category named Children Books

# Identify custom tag_names with children
pattern_for_Children = '\w*[-]?[cC]hildren\w*'
custom_tags_for_Children = identify_custom_categories(tag_table.tag_name, str(pattern_for_Children))

# Identify custom tag_names with childhood
pattern_for_Childhood = '\w*[-]?[cC]hildhood\w*'
custom_tags_for_Childhood = identify_custom_categories(tag_table.tag_name, str(pattern_for_Childhood))

# Identify custom tag_names with kids
pattern_for_Kids = '\w*[-]?[kK]id\w*'
custom_tags_for_Kids = identify_custom_categories(tag_table.tag_name, str(pattern_for_Kids))

custom_tags_for_Childre_and_Kids = custom_tags_for_Kids + custom_tags_for_Children + custom_tags_for_Childhood

print(custom_tags_for_Childre_and_Kids)

# Rename the custom tag_names as Children Books
preferred_name = 'Children Books'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_Childre_and_Kids
, preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append('Children Books')
KW_CategorY_Repos ['Possible Search KW'].append(custom_tags_for_Childre_and_Kids)

tag_table.tag_name.nunique()

#Check if Replacement Worked
tag_table.tag_name [50:100]

## Tag for Owned Books

## Tag for Owned Books
pattern_for_Owned = '\w*[-]?\w*[-]?[oO]wn\w*'
custom_tags_for_Owned = identify_custom_categories(tag_table.tag_name, str(pattern_for_Owned))
custom_tags_for_Owned 

# Rename the custom tag_names as Owned Books
preferred_name = 'Owned Books'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_Owned
, preferred_name))

#Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

## Find All Young Adult Books ( Fiction & Non - Fiction & Other Categories)

## Tag for Young Adult Books
pattern_for_YA = '\w*[-]?[Yy]oung\w*'
searchfor = ['ya', 'juvenile','teen']
custom_tags_for_YA1 = list(tag_table.tag_name[tag_table.tag_name.str.contains('|'.join(searchfor))])
custom_tags_for_YA2 = identify_custom_categories(tag_table.tag_name, str(pattern_for_YA))
custom_tags_for_YA = custom_tags_for_YA1 + custom_tags_for_YA2

> It looks like there is a borad range of you adult tags with different subcategories. Lets try to group them. First, lets get all the unique tags for young adult books.

set(custom_tags_for_YA)

#Lets group different young adult tags
Custom_Tag_YoungAdult =  [x for x in custom_tags_for_YA if x in ['ya','ya-books','ya-lit', 'young-adult','youngadult', 'juvenile', 'ya-contemporary','teen']]
Custom_Tag_YoungAdult_Fantasy =  [x for x in custom_tags_for_YA if x in ['ya-fantasy','young-adult-fantasy']]
Custom_Tag_YoungAdult_Fiction =  [x for x in custom_tags_for_YA if x in ['ya-fiction','young-adult-fiction','juvenile-fiction', 'teen-fiction']]
Custom_Tag_YoungAdult_Romance =  [x for x in custom_tags_for_YA if x in ['ya-romance']]
Custom_Tag_YoungAdult_Paranormal =  [x for x in custom_tags_for_YA if x in ['ya-paranormal']]
print('Number of tags in each category is as follows -',
      'YoungAdult:', len(Custom_Tag_YoungAdult),
      'YoungAdult_Fantasy:', len(Custom_Tag_YoungAdult_Fantasy),
      'YoungAdult_Fiction:', len(Custom_Tag_YoungAdult_Fiction),
      'YoungAdult_Romance:', len(Custom_Tag_YoungAdult_Romance),
      'YoungAdult_Paranormal:', len(Custom_Tag_YoungAdult_Paranormal))
                                                              

# Rename the custom tag_names for Young Adult Books
preferred_name = 'Young-Adult'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_YoungAdult,preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append( 'Young-Adult')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_YoungAdult)

preferred_name = 'Young-Adult-Fantasy'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_YoungAdult_Fantasy
, preferred_name))

preferred_name = 'Young-Adult-Fiction'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_YoungAdult_Fiction
, preferred_name))

preferred_name = 'Young-Adult-Romance'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_YoungAdult_Romance
, preferred_name))

preferred_name = 'Young-Adult-Paranormal'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_YoungAdult_Paranormal
, preferred_name))


# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append('Young-Adult-Fantasy')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_YoungAdult_Fantasy)

KW_CategorY_Repos ['Category'].append('Young-Adult-Fiction')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_YoungAdult_Fiction)

KW_CategorY_Repos ['Category'].append('Young-Adult-Romance')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_YoungAdult_Romance)

KW_CategorY_Repos ['Category'].append('Young-Adult-Paranormal')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_YoungAdult_Paranormal)

#Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

## Tag for Adult Readers

> Lets take a look at different segment of the 'tag_table' dataframe and then do a search to identify some tag_names used by adult readers

# Identify custom tag_names with the word fiction
pattern_for_Fiction = '\w*[-]?[fF]iction\w*'
custom_tags_for_Fiction = identify_custom_categories(tag_table.tag_name, str(pattern_for_Fiction))


# Identify some keywords that stood out while observing the dataframe and then do a search
searchfor = ['history', 'History','sci','mystery','Mystery','crime','Crime','Women','women','feminism','girl']
custom_tags_for_other_Adult_Books = tag_table.tag_name[tag_table.tag_name.str.contains('|'.join(searchfor))]
custom_tags_for_Adult_Books = list(custom_tags_for_other_Adult_Books) + custom_tags_for_Fiction
set(custom_tags_for_Adult_Books)

### Lets group tags relevant to 'fiction' and 'non-fictions'

# Identify custom tag_names for Adult Fiction and Non Fiction
pattern_for_Fiction = '\w*[-]?[fF]iction\w*'
custom_tags_for_Fiction = identify_custom_categories(tag_table.tag_name, str(pattern_for_Fiction))
custom_tags_for_Fiction

#Lets group the fiction and non fiction books
Custom_Tag_Fiction =  [x for x in custom_tags_for_Adult_Books if x in ['fiction','contemporary-fiction','adult-fiction',
                                                              'speculative-fiction','realistic-fiction', 
                                                              'general-fiction',  'literary-fiction','fiction-to-read', 
                                                              'modern-fiction',
                                                             'fiction-general', 'classic-fiction']]

Custom_Tag_NonFiction =  [x for x in custom_tags_for_Adult_Books if x in ['nonfiction','non-fiction','non-fiction-to-read']]
Custom_Tag_Historical_Fiction =  [x for x in custom_tags_for_Adult_Books if x in ['historical-fiction','fiction-historical',
                                                                                  'historic-fiction']]                                                      

# Rename the custom tag_names for Fictio and Non Fiction Books
preferred_name = 'Fiction'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_Fiction
, preferred_name))

preferred_name = 'Non-Fiction'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_NonFiction
, preferred_name))

preferred_name = 'Historical Fiction'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_Historical_Fiction
, preferred_name))


# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append( 'Fiction')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_Fiction)

KW_CategorY_Repos ['Category'].append( 'Non-Fiction')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_NonFiction)

KW_CategorY_Repos ['Category'].append( 'Historical Fiction')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_Historical_Fiction)

#Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

### Lets group tags for Audiobooks and Ebooks

# Identify custom tag_names Audiobooks
pattern_for_Audio_Books = '\w*[-]?[Aa]udi\w*'
custom_tags_for_Audio_Books = identify_custom_categories(tag_table.tag_name, str(pattern_for_Audio_Books))
custom_tags_for_Audio_Books

# Rename the custom tags for Audio Books
preferred_name = 'Audio Books'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_Audio_Books
, preferred_name))

# Identify custom tag_names Ebooks
pattern_for_Books = '\w*[-]?[bB]ook\w*'
custom_tags_for_Books = identify_custom_categories(tag_table.tag_name, str(pattern_for_Books))
set(custom_tags_for_Books )

# Identify custom tag_names Ebooks
searchfor = ['e-books', 'ebooks','ebook','e-book','kindle-books','kindle']
custom_tags_for_Ebooks = list(tag_table.tag_name[tag_table.tag_name.str.contains('|'.join(searchfor))])
set(custom_tags_for_Ebooks)

# Rename the custom tags for Ebooks
preferred_name = 'Ebooks'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_Ebooks
,preferred_name))

#Lets check the tag_names after applying the filter 
tag_table.tag_name[650:750]

### Lets Group Tags for Books Read in a Particular Year 

# Identify custom tag_names 4digits
pattern_for_4Digits = '\w*[-]?\w*[-]?\d{4}'
custom_tags_for_4Digits = identify_custom_categories(tag_table.tag_name, str(pattern_for_4Digits))
set(custom_tags_for_4Digits)

# Identify custom tag_names for books read in a particular year/time
pattern_for_Year = '\w*[-]?\w*[-]?[2]\d{3}'
custom_tags_for_Year = identify_custom_categories(tag_table.tag_name, str(pattern_for_Year))
custom_tags_for_Year.append('1990s')

# Rename the custom tags
preferred_name = 'Books Read By Year'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_Year
, preferred_name))

#Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

 ### Lets group other popular  tag_names related to Science, History,  Books for Women, Mystery & Crime,Science & Fantasy

#### Lets group tags relevant to 'Science'

#Lets group books on science
Custom_Tag_Science = identify_custom_categories(tag_table.tag_name, str('[Ss]cience$'))
Custom_Tag_Science

#Lets rename books on science
preferred_name = 'Science'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_Science
, preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append( 'Science')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_Science)

#Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

#### Lets group tags relevant to 'History'

#Lets group books on History
Custom_Tag_History = identify_custom_categories(tag_table.tag_name, str('[Hh]istory\w*'))
Custom_Tag_History

#Lets rename books on History
preferred_name = 'History'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,Custom_Tag_History
, preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append( 'History')
KW_CategorY_Repos ['Possible Search KW'].append(Custom_Tag_History)

#Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

> It seems that tags for History and Science related books are well defined by all the users, they did not choose any other name to shelve these books.

#### Lets group tags relevant to 'Women'


# Identifythe custom tags
searchfor = ['Women','women','feminism']
custom_tags_for__Women_Books = list(tag_table.tag_name[tag_table.tag_name.str.contains('|'.join(searchfor))])

# Rename custom tags
preferred_name = 'Women Book List'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for__Women_Books
, preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append( 'Women Book List')
KW_CategorY_Repos ['Possible Search KW'].append(custom_tags_for__Women_Books)

#### Lets group tags relevant to 'Crime & Mystery'

#Lets group books on Crime and Mystery
searchfor = ['[mM]ystery$', '[cC]rime$']
custom_tags_for_Mystery_Crime = list(tag_table.tag_name[tag_table.tag_name.str.contains('|'.join(searchfor))])
set(custom_tags_for_Mystery_Crime)

# Lets Rename
preferred_name = 'Crime & Mystery'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_Mystery_Crime
, preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append(  'Crime & Mystery')
KW_CategorY_Repos ['Possible Search KW'].append(custom_tags_for_Mystery_Crime)

# Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

#### Lets group tags relevant to 'Science Fiction & Fantasy'

# Lets group books on Science & Fantasy
searchfor = ['\w*[-]?[sS]ci(?!.*ence)\w*', '^\w*(?!.*Young-Adult).[fF]antasy$',
 'science-fiction'] #(?!.*Word) will not contain the word
custom_tags_for_Science_Fantasy = list(tag_table.tag_name[tag_table.tag_name.str.contains('|'.join(searchfor))])

set(custom_tags_for_Science_Fantasy)

# Lets Rename
preferred_name = 'Science Fiction & Fantasy'
tag_table.tag_name = tag_table.tag_name.apply(lambda x: replace_custom_categories (x,custom_tags_for_Science_Fantasy
, preferred_name))

# Save all the names for favorites in a dictionary to help keyword search later in the project
KW_CategorY_Repos ['Category'].append('Science Fiction & Fantasy')
KW_CategorY_Repos ['Possible Search KW'].append(custom_tags_for_Science_Fantasy)

#Lets check the tag_names after applying the filter 
tag_table.tag_name.nunique()

> Now, we can group the dataframe by tag_name to get an idea of how frequently users like to use such categories

tag_table_shortened = tag_table[['tag_id','tag_name','goodreads_book_id']]
Frequency_per_tag = tag_table_shortened.groupby(['tag_name']).count().sort_values(by='goodreads_book_id',ascending = False)
Frequency_per_tag.rename(columns = {'goodreads_book_id':'Frequency'}, inplace = True)
Frequency_per_tag.head(10)

> Lets get an idea of number of books per catgory. When we merged different Tags to one single category, the tag count for a give book in each category got added. It means if a book was tagged as 'Fav' and 'myFav', it will have a tag frequency of  when the tags are merged to a single tag called Favorites. So we will have to drop duplicates to idetify books per tag

tag_table_shortened = tag_table[['tag_name','goodreads_book_id']]
tag_table_shortened.loc[tag_table_shortened.duplicated (keep = False),:]
books_per_tag = tag_table_shortened.drop_duplicates()
books_per_tag.head(20)

tag_table.head()

# Make a dataframe for category keyword search
KW_Repository = pd.DataFrame.from_dict(KW_CategorY_Repos)
KW_Repository.head()

# Export Tidy Dataset to CSV for EDA

By using data wrangling we have reduced the tag_ids from 32963 to 348. Lets import the clean dataset to a CSV format for exploratory data analysis.

KW_Repository.to_csv('Category_KW_Respository.csv', encoding = 'utf-8')
tag_table.to_csv('Tidy_Tag_Table.csv', encoding = 'utf-8')
Frequency_per_tag.to_csv('Tidy_Data_for_Tag_Frequency.csv', encoding = 'utf-8')
books_per_tag.to_csv('Tidy_Data_for_Books_Per_Tag.csv',encoding = 'utf-8')

## Few Comments

- Users like to have sub categories within a category [such as historical fiction, children - nonfiction, young-adult-fiction]]. We can use this to provide some built in tags
- Users also prefer to shelve for specific series or Author (jk rowling, harry potter is the most frequent)
- After defining all the categories, we can see what category got most 5 stars
- For highly rated books, we can  look at curtomized tag names within a given category, and then check whats the most popular type of book that they like (may be young adult rated fiction better than than historical - no fiction etc]
- Whats the age group of different readers (count the books in adult, kids and YA category)
